added also hiseqX_run_times
